# Add Variant class to optimise URL generation performance in request