# Description

This builds on changes from a separate PR: https://github.com/imgix/imgix-rb/pull/103

Rather than escape the same query string on each request, it can be pre-generated on application start. This removes any need to escape query strings in requests. A similar approach could be taken for `to_srcset` with a `SrcSet` class.

An application might be defining Imgix options in a hash like this:

```ruby
IMGIX_OPTIONS = {
  type_1: { auto: 'compress,format', fit: 'crop', crop: 'top', w: 200, h: 200 },
  type_2: { auto: 'compress,format', fit: 'crop', crop: 'top', w: 400, h: 400 },
}
```

The application could change the definitions to:

```ruby
IMGIX_OPTIONS = {
  type_1: Imgix::Variant.new(auto: 'compress,format', fit: 'crop', crop: 'top', w: 200, h: 200),
  type_2: Imgix::Variant.new(auto: 'compress,format', fit: 'crop', crop: 'top', w: 400, h: 400),
}
```

And then pass the variant to `Imgix::Path#to_url`:

```ruby
client.path('/image.jpg').to_url(variant: IMGIX_OPTIONS[:type_1])
```

Bench output:

```
❯ bundle exec ruby script/bench_to_url.rb
Warming up --------------------------------------
  Imgix::Path#to_url     2.772k i/100ms
Imgix::Path#to_url with variant
                        25.965k i/100ms
Calculating -------------------------------------
  Imgix::Path#to_url     28.021k (± 0.9%) i/s -    141.372k in   5.045620s
Imgix::Path#to_url with variant
                        262.187k (± 1.2%) i/s -      1.324M in   5.051381s

Comparison:
Imgix::Path#to_url with variant:   262186.7 i/s
  Imgix::Path#to_url:    28021.2 i/s - 9.36x  (± 0.00) slower
```

## Checklist: New Feature

- [ ] Each commit follows the [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/#summary) format: e.g. `chore(readme): fixed typo`. See the end of this file for more information.
- [ ] Any breaking changes are specified on the commit on which they are introduced with `BREAKING CHANGE` in the body of the commit.
- [ ] If this is a big feature with breaking changes, consider [opening an issue](https://github.com/imgix/imgix-rb/issues/new?labels=&template=feature_request.md&title=) to discuss first. This is completely up to you, but please keep in mind that your PR might not be accepted.
- [ ] Run unit tests to ensure all existing tests are still passing
- [ ] Add new passing unit tests to cover the code introduced by your PR
- [ ] Update the readme
- [ ] Add some [steps](#steps-to-test) so we can test your cool new feature!