@sherwinski 2021-03-30
Hey @stevehodgkiss, thanks for opening for the PR. I'm interested in reviewing this but it looks like there's a file missing (`imgix/variant`) that still needs to be checked in. That is unless this PR is still a WIP, in which case feel free to ping me when you're ready 👍