@stevehodgkiss 2021-03-31
Hi @sherwinski, whoops, that's fixed now. Ready for review - interested in your thoughts on the change. A similar approach could be taken to improve `to_srcset` performance.