@sherwinski 2021-04-28
Hey @stevehodgkiss,
First off, I wanted to say thanks again for opening this PR and putting so much thought into improving this library in this way. Our team is always looking for ways to improve performance as well as usability of our SDKs; and this PR has pushed us to think more deeply about how else we can prevent repeat work. Having said that, we are opting to not accept this change at this time, and I’d like to explain why.
After discussing it further with the team, we agreed that this is a feature that should be implemented across our [entire kit](https://docs.imgix.com/libraries#client-libraries) of open source libraries, spanning approximately 10 different languages. As a result, this requires extra work before release: ensuring cross-kit consistency, considering language-specific idioms, and investigating for potential edge cases. For this reason, we’d like to put more time into designing this feature internally before implementing it across the board.